GStools/deployment/GStools/mcp_output/requirements.txt
fastmcp
fastapi
uvicorn[standard]
pydantic>=2.0.0
gstools-cython>=1,<2
emcee>=3.0.0
hankel>=1.0.0
meshio>=5.1.0
numpy>=1.20.0
pyevtk>=1.1.1
scipy>=1.1.0
